Added EditFamily and Promotion blocks
Added Edit Family block that allows additional configurations for what is shows and required.  Also support adding a check-in group during registration for Already Belongs check-in.
Added Group Promotion block to allow Already Belongs groups to be promoted from one age/grade to another.  Also supports backing up groups prior to promotion and undoing a promotion if not configured correctly.
Added Age Promotion block.  This block allows Already Belongs groups to be synchronized to birth dates when groups need to be reconfigured to balance classrooms.  It also supports backups and an undo promotion if groups not configured correctly.