Initial Creation
Created a block which outputs lines for a SQL script calling a stored procedure and a SQL script to create the stored procedure.   This tool will allow you to copy groups from one type to another and map roles and attributes to the new type.  It also optionally copies attendance.  The Delete Option is not implemented.  Running multiple times should be supported and update the information, however members and attributes that are deleted will not be deleted on subsequent runs.