Initial Version of Metrics Plug-in
This plug-in includes a rest controller, blocks and lava format templates for creating metric dashboards.  There are lava templates for a block, vertical and horizontal versions of the display that are responsive to the block settings for what to show.

There may be some bugs - final testing is not yet complete.