Transnational Plugin
This is the 10.3 Plugin Files that I did for the Transnational Plugin.  They have stopped working in v11.

The basic function of the plugin is to choose a gateway, a start date, and an end date and use the same code of the download payments block to pull transactions from Transnationa, fitler only the ones that are settled, group them into batches and for each batch, show a  fund summary on one pill, and the list of transactions on the second.

I've included a screenshot of the current views in the assets folder of this branch.
Three Views
Initial View(Settlement-Batch-Fund_Summary-DefaultView) shows the deposit details with the words deposit/withdrawal changing to match its value  The confirm deposit button looks through the transactions and updates the Reconciled On Value
Transaction List(Transaction-View to Show ): Lists the Rock transactions that make up that settlement deposit.
Include in Print(After-Deposit-Confirmed): When all of the transactions in a settlement batch have been confirmed and the reconciled on date set to the date the Confirm Deposit button is clicked, the confirm deposit button disappears and there is a check box for including an item in print.  This lets the user, if they want to print the fund summaries, pick which they want to include.  In the lava, there is a check for the total number confirmed and if it matches the number of transactions, the deposit is considered confirmed.